requirements.txt
Django==4.0.1
Markdown==3.1.1
djangorestframework==3.12.4
django-environ==0.8.1
drf-nested-routers==0.93.4
djangorestframework-simplejwt==5.1.0